src/dcl_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dcl-agent"
version = "0.1.0"
description = "Agentic Runtime for executing Domain Context Language (DCL) instructions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lark>=1.1.9",
    "google-genai>=0.1.0",
    "PyYAML>=6.0",
]

[tool.setuptools]
packages = ["dcl_agent"]

[tool.setuptools.package-dir]
dcl_agent = "."
